修改商品和订单接口的bug/新增购物车和收藏的业务代码框架

+++ src/main/java/com/shopping/inferior/management/orders/service/OrderServiceImp.java
@@ -32,9 +32,10 @@ public class OrderServiceImp implements OrderService{
         return 1;
     }
 
-    public int deleteOrders(Integer id,Integer goodsId){
+    public int deleteOrders(Integer id,Integer goodsId,Integer status){
         try{
             Orders orders = ordersMapper.selectByPrimaryKey(id);
+            if(orders.getOrderStatus() > status)return 0;
             ordersMapper.deleteByPrimaryKey(id);
             Goods goods = goodsMapper.selectByPrimaryKey(goodsId);
             if(goods.getGoodsStatus() == 0)goods.setGoodsStatus(1);
